tests/krate/publish: Extract `emails` module
